rename: 디렉토리 구조 변경에 따른 경로 재설정
- common 구조 변경에 따른 컴포넌트 경로들 재설정
