+++ fe/src/components/mainPage/issueTable/IssueTable.jsx
@@ -22,5 +22,3 @@ const IssueTable = ({ issueState, issuePageData, onChangeIssueState }) => {
 };
 
 export default IssueTable;
-
-
